poly/line, all prototypes are 1 deep now, segments inherit Array.prototype
