+++ src/primitives/lines/line.js
@@ -1,5 +1,4 @@
 import Constructors from "../constructors";
-import LinePrototype from "../prototypes/line";
 import { resize } from "../../arguments/resize";
 import { get_line } from "../../arguments/get";
 import {
@@ -8,13 +7,16 @@ import {
 } from "../../arguments/functions";
 import {
   add,
-  scale
+  scale,
 } from "../../core/algebra";
 import Static from "./static";
+import methods from "./methods";
+
+// LineProto.prototype.constructor = LineProto;
 
 export default {
   line: {
-    P: LinePrototype.prototype,
+    P: Object.prototype,
 
     A: function () {
       const l = get_line(...arguments);
@@ -24,10 +26,15 @@ export default {
     },
 
     G: {
-      length: () => Infinity,
+      // length: () => Infinity,
+      dimension: function () {
+        return [this.vector, this.origin]
+          .map(p => p.length)
+          .reduce((a, b) => Math.max(a, b), 0);
+      },
     },
 
-    M: {
+    M: Object.assign({}, methods, {
       inclusive: function () { this.domain_function = include_l; return this; },
       exclusive: function () { this.domain_function = exclude_l; return this; },
       clip_function: dist => dist,
@@ -36,7 +43,7 @@ export default {
         const end = scale(this.vector, length);
         return `M${start[0]} ${start[1]}l${end[0]} ${end[1]}`;
       },
-    },
+    }),
 
     S: Static
