+++ src/primitives/polygon/index.js
@@ -1,17 +1,16 @@
 import Constructors from "../constructors";
-import Prototype from "../prototypes/polygon";
+import methods from "../shared/polygon";
 import { semi_flatten_arrays } from "../../arguments/resize";
 import { include, exclude } from "../../arguments/functions";
 import { subtract } from "../../core/algebra";
 import {
   convex_hull,
   make_regular_polygon,
-  // straight_skeleton,
 } from "../../core/geometry";
 
 export default {
   polygon: {
-    P: Prototype.prototype,
+    P: Array.prototype,
     A: function () {
       this.push(...semi_flatten_arrays(arguments));
       // this.points = semi_flatten_arrays(arguments);
@@ -21,7 +20,7 @@ export default {
         // .map(ps => Constructors.segment(ps[0][0], ps[0][1], ps[1][0], ps[1][1]));
       this.vectors = this.sides.map(side => subtract(side[1], side[0]));
       // this.sectors
-      // Object.defineProperty(this, "domain_function", { writable: true, value: exclude });
+      Object.defineProperty(this, "domain_function", { writable: true, value: include });
     },
     G: {
       // todo: convex test
@@ -35,16 +34,13 @@ export default {
       //   return this.sides;
       // },
     },
-    M: {
+    M: Object.assign({}, methods, {
       inclusive: function () { this.domain_function = include; return this; },
       exclusive: function () { this.domain_function = exclude; return this; },
       segments: function () {
         return this.sides;
       },
-      // straightSkeleton: function () {
-      //   return straight_skeleton(this);
-      // },
-    },
+    }),
     S: {
       fromPoints: function () {
         return this.constructor(...arguments);
