+++ src/primitives/vector/args.js
@@ -1,7 +1,7 @@
 import { get_vector } from "../../arguments/get";
 
 const VectorArgs = function () {
-  get_vector(arguments).forEach(n => this.push(n));
+  this.push(...get_vector(arguments));
 };
 
 export default VectorArgs;
